Possible update for readme for Mac situation
I prefer to add things to my path than create a `$HOME/bin` directory that Mac doesn't really 'do', at least by default.  This workflow worked for me.  If you think this stinks, it would be nice to have something mentioning the profile or bashrc, which however on Mac doesn't seem to have the same role as on Linux.